backend/QVtoolapi/requirements.txt
Django==3.1.2
django_guardian==2.3.0
djangorestframework==3.12.2
django-cors-headers==3.3.0
djrill==2.1.0
gunicorn==20.0.4
mailchimp3==3.0.14
premailer==3.7.0
psycopg2-binary==2.8.6
sendgrid==6.3.1
six==1.15.0
whitenoise==5.2.0
